Added size limit

+++ FinalProject.cpp
@@ -20,32 +20,32 @@ enum sort;
 /*
  1. Function prototype:
     void checkSort(int size, int sort);
- 
+
  2. Function description:
     This function is by far the largest, and is the only function called from the driver.
-    It takes in a number for the array size the user wants sorted, and a number that 
+    It takes in a number for the array size the user wants sorted, and a number that
     corresponds to which sort is to be implemented. To make the program more legible, the
-    input number corresponds with the function name as defined in "enum sort" in the 
+    input number corresponds with the function name as defined in "enum sort" in the
     header file.
- 
+
     Three arrays and counters are created. The arrays are of the input size, and the counters
     correspond to each array. The first array is built to be a reverse order array using a
     helper function. The second is an in order array. The third is a random order array.
- 
+
     The desired sorting function is then applied to each array. The three counters are passed
     into the functions to give a general measure of the efficiency of the sort.
- 
-    Then, for each array, the function prints whether the array was properly sorted 
-    (a helper function checks), the count from that sort is printed, and the sorted 
+
+    Then, for each array, the function prints whether the array was properly sorted
+    (a helper function checks), the count from that sort is printed, and the sorted
     array is printed (if the size exceeds 10, ellipses are used and the first 5 and
     last 5 elements are printed).
- 
+
  3. Example:
     checkSort(size, userInput); //size is array size, userInput is the desired sort
- 
+
  4. Precondition:
         Size and userInput given by the user are passed into the function.
- 
+
     Postcondition:
         Neither size nor userInput are changed. The function prints the gathered info, and
         deletes everything created inside the function.
@@ -61,7 +61,7 @@ void checkSort(int size, int sort)
     createReverseOrderList(array, size);
     createInOrderList(secondArray, size);
     createRandomList(thirdArray, size);
-    
+
     cout << endl;
     if (sort == HEAPSORT)
     {
@@ -146,7 +146,7 @@ void checkSort(int size, int sort)
         cout << endl;
     }
     cout << endl;
-    
+
     //secondArray is the reverse order array
     if (!isSorted(secondArray, size))
     {
@@ -180,7 +180,7 @@ void checkSort(int size, int sort)
         cout << endl;
     }
     cout << endl;
-    
+
     //thirdArray is random order array
     if (!isSorted(thirdArray, size))
     {
@@ -214,7 +214,7 @@ void checkSort(int size, int sort)
         cout << endl;
     }
     cout << endl;
-    
+
     //free memory
     delete [] array;
     delete [] secondArray;
@@ -224,17 +224,17 @@ void checkSort(int size, int sort)
 /*
  1. Function prototype:
     void callHeapSort(int *arr, int size, int &counter);
- 
+
  2. Function description:
-    This function creates a heap, calls heapsort, and sets the counter equal to the 
+    This function creates a heap, calls heapsort, and sets the counter equal to the
     counter in the class. The input array is sorted once this function finishes.
- 
+
  3. Example:
     callHeapSort(array, size, counter); //This is only ever called in checkSort.
- 
+
  4. Precondition:
         An unsorted array, its size, and a counter = 0 are entered.
- 
+
     Postcondition:
         The array is sorted, its size is unchanged, and the counter has been added to
         so that it roughly measures how long the sort took.
@@ -249,17 +249,17 @@ void callHeapSort(int *arr, int size, int &counter)
 /*
  1. Function prototype:
     void callTreeSort(int *arr, int size, int &counter);
- 
+
  2. Function description:
     This function creates a Tree from the input array, sorts the tree and sets the array
     to the sort tree function (which returns an array), and gets the counts from the sort.
- 
+
  3. Example:
     callTreeSort(array, size, counter); //This is only ever called in checkSort.
- 
+
  4. Precondition:
         An unsorted array, its size, and a counter = 0 are entered.
- 
+
     Postcondition:
         The array is sorted, its size is unchanged, and the counter has been added to
         so that it roughly measures how long the sort took.
@@ -274,17 +274,17 @@ void callTreeSort(int *arr, int size, int &counter)
 /*
  1. Function prototype:
     void callMergeSort(int *arr, int size, int &counter);
- 
+
  2. Function description:
     This function calls merge sort from the info provided by the user.
     The array is sorted after this function is called.
- 
+
  3. Example:
     callMergeSort(array, size, counter); //This is only ever called in checkSort
- 
+
  4. Precondition:
         An unsorted array, its size, and a counter = 0 are entered.
- 
+
     Postcondition:
         The array is sorted, its size is unchanged, and the counter has been added to
         so that it roughly measures how long the sort took.
@@ -297,17 +297,17 @@ void callMergeSort(int *arr, int size, int &counter)
 /*
  1. Function prototype:
     void callQuickSort(int *arr, int size, int &counter);
- 
+
  2. Function description:
     This function calls quick sort from the info provided by the user.
     The array is sorted after this function is called.
- 
+
  3. Example:
     callQuickSort(array, size, counter); //This is only ever called in CheckSort
- 
+
  4. Precondition:
         An unsorted array, its size, and a counter = 0 are entered.
- 
+
     Postcondition:
         The array is sorted, its size is unchanged, and the counter has been added to
         so that it roughly measures how long the sort took.
@@ -343,7 +343,7 @@ void createRandomList(int *arr, int size)
 {
     for (int i = 0; i < size; i++)
     {
-        double randomNum = rand() % size; //creates numbers between 0 and the array size
+        int randomNum = rand() % size; //creates numbers between 0 and the array size
         arr[i] = randomNum;
     }
 }
@@ -351,7 +351,7 @@ void createRandomList(int *arr, int size)
 /*
  * This is another simple helper function. I didn't think it was big enough to merit
  * the whole specified comment block applied to significant functions. It checks if an
- * input array of input size is sorted. It starts by assuming the array is sorted, and 
+ * input array of input size is sorted. It starts by assuming the array is sorted, and
  * if any element is smaller than the element before it, the boolean is switched to false.
  * The test fails if all elements in the array are the same. That is why I also print the
  * sorted array in the checkSort function. This is only called in checkSort.
@@ -368,3 +368,8 @@ bool isSorted(int *arr, int size)
     }
     return sorted;
 }
+
+
+
+
+
